
\name{cardatdbtools-package}
\alias{cardatdbtools-package}
\alias{cardatdbtools}
\docType{package}
\title{
getPassword and connect2postgres
}
\description{
getPassword and connect2postgres~
}
\details{
\tabular{ll}{
Package: \tab cardatdbtools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-08-31\cr
License: \tab GPL2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <ivan.hanigan@gmail.com>

}
\references{

}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
